requirements.txt
# Core LLM and RAG Stack
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
sentence-transformers
transformers
faiss-cpu  
huggingface-hub

# Ollama LLM Integration
langchain-ollama>=0.1.0
ollama>=0.1.7

# Document loaders
pypdf
pdfminer.six>=20221105

# Text preprocessing
tqdm>=4.66.0

# Web app interface
streamlit>=1.35.0
streamlit-option-menu>=0.3.6

# Logging & file handling
python-dotenv>=1.0.1
watchdog>=4.0.0

# Optional: For file encoding handling (future-proofing)
chardet>=5.2.0

# google
google.generativeai
langchain-google-genai